Development Log Entry 3 Date: Tuesday November 25th, 2025 Lesson number: 4 1. Goal for today (C i)
Today I planned to implement my validation effectivly using JavaScript and CSS, so that I never had to do any validation with python. This is so that I can flash messages directly underneath the text boxes. I planned to get it all completed today.

2. Work completed (C iii)

I was able to fully validate the name (no spaces etc...) and the date of burth (not in the future), and all required fields. I was able to display the error messages under the field. I still need to flash them on top and I will do so using the same JavaScript method.

3. Did you follow your plan? (C iii and C iv)
Yes:
I completed all of my validation and I only need to add the box at the top which I could not do since I ran out of time.
(When you complete this section, you are providing evidence for Criterion C iv.)
4. Next step (C i)

Next lesson I need to fix glitches with my JSON and display and add the error box at the top of my form.