Development Log Entry 6 Date: Tuesday December 2nd, 2025 Lesson number: 6 1. Goal for today (C i)
Today I had planned to debug my code, test it further, and add a delete functionality and possibly an edit functionality which let me change the JSON.

2. Work completed (C iii)

I was able to add a delete button which successfully deletes my work which uses the same CSS logic as my Submit button for aesthetics, and I was able to fully comment my python code in order for it to be usable in my reflection.
3. Did you follow your plan? (C iii and C iv)
Yes:
Yes, except I did not have time to add edit functionality and I did not know how to implement it so I deemed it too risky to do as it may mess up my whole code logic.

(When you complete this section, you are providing evidence for Criterion C iv.)
4. Next step (C i)
I'm finished !! :)